pyproject.toml
[tool.poetry]
name = "ecom-bot"
version = "0.1.0"
description = ""
authors = ["Mikhail Noskov <noskovmm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.2.1"
openai = "^2.7.1"
langchain = "^1.0.4"
langchain-openai = "^1.0.2"
langgraph = "^1.0.2"
python-json-logger = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
